feat: enhance Content component with fullscreen support and YTDLiveChat integration; refactor YTDLiveChat for improved rendering; add useIsFullscreen hook
