+++ src/content/features/Draggable/styles/Draggable.module.css
@@ -5,6 +5,4 @@
   left: 0px;
   width: 100%;
   height: 100vh;
-  pointer-events: none;
-  z-index: 1000000;
 }
