+++ src/content/hooks/useIsFullscreen.ts
@@ -1,0 +1,23 @@
+import { useEffect, useState } from 'react'
+
+export const useIsFullScreen = () => {
+  const [isFullscreen, setIsFullscreen] = useState(false)
+
+  useEffect(() => {
+    const handleFullscreenChange = () => {
+      if (document.fullscreenElement !== null) {
+        setIsFullscreen(true)
+      } else {
+        setIsFullscreen(false)
+      }
+    }
+
+    document.addEventListener('fullscreenchange', handleFullscreenChange)
+
+    return () => {
+      document.removeEventListener('fullscreenchange', handleFullscreenChange)
+    }
+  }, [])
+
+  return isFullscreen
+}
