+++ src/content/styles/index.css
@@ -1,16 +1,11 @@
 /* --------------------------- fullscreen setting --------------------------- */
-ytd-app {
-  right: 0 !important;
-  scrollbar-width: none;
-}
-
 .extension-root-d774ba85-ed7c-42a2-bf6f-a74e8d8605ec {
   position: fixed;
   top: 0;
   right: 0;
   width: 100%;
   height: 100%;
-  z-index: 1000000;
+  z-index: 1000;
   pointer-events: none;
 }
